Flexprobe Encryption Detection update
The encryption detection data can now be used for classification. The sample
is passed to the classification tool via ZMQ before the flow data is exported.
If ZMQ is not present, this feature will not be available. In case no tool
capable of processing the statistical sample is running, the detection will
only use simple checks for presence of already known protocols.