Improved description of output
Based on feedback from #81, it was not very clear what is the output of ipfixprobe when some output plugin is used.

The aim of this patch is to make the readme file more general - if a user enables some processing plugin (by `-p`), there are extended fields included in the output flow records. It works for `unirec`, `ipfix`, or `text` plugins.