plugin: input and process plugin for unix domain socket
Systems were ipfix traffic is metered/counted by the
hardware support this new input plugin sock and the
corresponding process plugin sockpktinfo can be useful
to receive and process ipfix flow information in form
of switch records via a unix domain socket which then
can be formatted in ipfix data packet and exported to
a collector via ipfixprobe.

Add some new fields for IPv4 and IPv6 that could be
useful.